# only call DubboBootstrap.reset(); once to make the test faster