## What is the purpose of the change
The propose of this change is to make tests run faster. `DubboBootstrap` is only used once in the test `org.apache.dubbo.generic.GenericServiceTest`. There are five tests in the test class `org.apache.dubbo.generic.GenericServiceTest`. However, only `testGenericCompatible` tries to modify with `DubboBootstrap` and other four tests are not dealing with `DubboBootstrap`. Repeatly calling `DubboBootstrap.reset();` may cause a big cost in time. On our local machine, the time can jump from 4.281 s to 1.429 s after applying the change.

## Brief changelog
Change the `@BeforeEach` annotation to `@AfterEach` annotation to only call `DubboBootstrap.reset();` once.

## Verifying this change
We confirm that if make this change, the test run time can jump from `4.281 s` to `1.429 s`.

<!-- Follow this checklist to help us incorporate your contribution quickly and easily: -->

## Checklist
- [x] Make sure there is a [GitHub_issue](https://github.com/apache/dubbo/issues) field for the change (usually before you start working on it). Trivial changes like typos do not require a GitHub issue. Your pull request should address just this issue, without pulling in other changes - one PR resolves one issue.
- [ ] Each commit in the pull request should have a meaningful subject line and body.
- [ ] Write a pull request description that is detailed enough to understand what the pull request does, how, and why.
- [ ] Check if is necessary to patch to Dubbo 3 if you are work on Dubbo 2.7
- [ ] Write necessary unit-test to verify your logic correction, more mock a little better when cross module dependency exist. If the new feature or significant change is committed, please remember to add sample in [dubbo samples](https://github.com/apache/dubbo-samples) project.
- [ ] Add some description to [dubbo-website](https://github.com/apache/dubbo-website) project if you are requesting to add a feature.
- [ ] GitHub Actions works fine on your own branch.
- [ ] If this contribution is large, please follow the [Software Donation Guide](https://github.com/apache/dubbo/wiki/Software-donation-guide).